pyproject.toml
[project]
name = "gitgrabber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "gitpython>=3.1.43",
    "requests>=2.32.3",
    "ruff>=0.8.4",
]
